Removed the option of setting the worker temp directory directly in the Runnable
The implementation was not aware that a worker can take on multiple
analyses. Being implemented at the Runnable level, there was also the risk
that the method would use job parameters to make the directory name,
unaware that the name is set for all jobs of all analyses the worker takes.

The impact is anyway minimal, as the only Ensembl Runnable that was using
this feature is a Compara Runnable, which can be ported to the new
resource-class based system.

And if Runnables still need to change the worker's temp directory, there
are still free to override worker_temp_directory() and
cleanup_worker_temp_directory().